Sync (#158)
* Correct discharge of battery

* Explicit Units

* removed middleware

* Migrated to typed OSRM queries

* Added distance to journey

* Journey and EV utility methods

* Working SoC i think, numbers makes sense. Sim can run to end

* Migrated to uniform SoC now it should be a value between 0-1.

* Integrated Price

* Parallel CreateEVs

* Made CandidateSearch Parallel again

* Fixed bug where station queue was always 0

* Temp for main

+++ Engine/Init/EngineSettings.cs
@@ -48,7 +48,7 @@ public class EngineSettings
     /// <summary>
     /// Gets the interval at which the engine checks for urgent EVs.
     /// </summary>
-    required public ushort BatteryIntervalForCheckUrgency { get; init; }
+    required public float BatteryIntervalForCheckUrgency { get; init; }
 
     /// <summary>
     /// Gets the number of seconds an EV is expected to spend charging at a station, used for simulating charging time and scheduling.
